Split redux into hooks, reducers and store
